+++ recipes/AlloST/ST/allost1/train_espnet.py
@@ -174,7 +174,8 @@ class ST(sb.core.Brain):
         if stage == sb.Stage.TRAIN:
             self.train_stats = stage_stats
         else:
-            stage_stats["ACC"] = self.acc_metric.summarize()
+            if stage == sb.Stage.VALID:
+                stage_stats["ACC"] = self.acc_metric.summarize()
             current_epoch = self.hparams.epoch_counter.current
             valid_search_interval = self.hparams.valid_search_interval
